Moved creational responsibilities into factory.
